requirements.txt
python-telegram-bot>=12.0.0
pymongo>=3.9.0
PyYAML>=5.1.2
flake8>=3.7.8
flake8-print
pre-commit